+++ compiler/testData/debug/stepping/ifThen.kt
@@ -1,0 +1,25 @@
+// FILE: test.kt
+
+fun foo() {
+    if (flag) {
+        return
+    }
+}
+
+var flag = true
+
+fun box() {
+    foo()
+    flag = false
+    foo()
+}
+
+// LINENUMBERS
+// test.kt:12 box
+// test.kt:4 foo
+// test.kt:5 foo
+// test.kt:13 box
+// test.kt:14 box
+// test.kt:4 foo
+// test.kt:7 foo
+// test.kt:15 box
