+++ compiler/testData/debug/stepping/multilineInfixCall.kt
@@ -1,0 +1,16 @@
+// FILE: test.kt
+
+fun box() {
+    1 foo
+         1
+}
+
+infix fun Int.foo(i: Int) {
+}
+
+// LINENUMBERS
+// test.kt:4 box
+// test.kt:5 box
+// test.kt:4 box
+// test.kt:9 foo
+// test.kt:6 box
