+++ compiler/testData/debug/stepping/primitiveNullChecks.kt
@@ -1,0 +1,12 @@
+// FILE: test.kt
+
+fun box(): String {
+    42!!
+    42.toLong()!!
+    return "OK"!!
+}
+
+// LINENUMBERS
+// test.kt:4 box
+// test.kt:5 box
+// test.kt:6 box
