--- compiler/testData/lineNumber/custom/beforeGotoToWhileStart.kt
@@ -1,13 +1,0 @@
-fun foo() {
-    while (true) {
-        if (testSome()) {
-            break
-        }
-    }
-}
-
-fun testSome(): Boolean {
-    return false
-}
-
-// 2 3 4 2 7 10
