--- compiler/testData/lineNumber/custom/compileTimeConstant.kt
@@ -1,6 +1,0 @@
-fun foo() {
-    val x =
-            42
-}
-
-// 2 +3 2 4
