--- compiler/testData/lineNumber/custom/functionCallWithLambdaParam.kt
@@ -1,15 +1,0 @@
-fun foo() {
-    foo({
-            val a = 1
-        })
-
-    foo() {
-        val a = 1
-    }
-}
-
-fun foo(f: () -> Unit) {
-    f()
-}
-
-// 2 6 9 12 13 3 4 7 8
