--- compiler/testData/lineNumber/custom/ifThen.kt
@@ -1,9 +1,0 @@
-fun foo() {
-    if (flag) {
-        return
-    }
-}
-
-val flag = true
-
-// 2 3 5 7 7
