--- compiler/testData/lineNumber/custom/ifThenElse.kt
@@ -1,17 +1,0 @@
-fun foo() {
-    if (flag) {
-        System.out?.println()
-    } else {
-        System.out?.println()
-    }
-    
-    val b = if (flag) {
-        System.out?.println()
-    } else {
-        System.out?.println()
-    }
-}
-
-val flag = true
-
-// 2 3 5 6 +8 9 11 8 13 15 15
