--- compiler/testData/lineNumber/custom/ifThenElseFalse.kt
@@ -1,10 +1,0 @@
-fun cond() = false
-
-fun foo() {
-    if (cond())
-        cond()
-    else
-         false
-}
-
-// 1 4 5 7 8
