--- compiler/testData/lineNumber/custom/multilineFunctionCall.kt
@@ -1,10 +1,0 @@
-fun foo() {
-    foo(
-            1 + 1
-    )
-}
-
-fun foo(i: Int) {
-}
-
-// 2 +3 2 5 8
