--- compiler/testData/lineNumber/custom/multilineInfixCall.kt
@@ -1,9 +1,0 @@
-fun foo() {
-    1 foo
-         1
-}
-
-infix fun Int.foo(i: Int) {
-}
-
-// 2 3 2 4 7
