--- compiler/testData/lineNumber/custom/primitiveNullChecks.kt
@@ -1,7 +1,0 @@
-fun box(): String {
-    42!!
-    42.toLong()!!
-    return "OK"!!
-}
-
-// 2 3 4
