--- compiler/testData/lineNumber/custom/smapInlineInIntrinsicArgument.kt
@@ -1,17 +1,0 @@
-fun box(){
-    test() +
-            fail()
-
-    fail() +
-                test()
-}
-
-inline fun test() : String {
-    return "123"
-}
-
-fun fail() : String {
-    throw AssertionError("fail")
-}
-
-// 2 18 3 5 6 19 7 10 14
