--- compiler/testData/lineNumber/custom/tryCatchExpression.kt
@@ -1,15 +1,0 @@
-fun foo() {
-    try {
-        System.out?.println()
-    } catch (e: Throwable) {
-        return
-    }
-    
-    val t = try {
-        System.out?.println()
-    } catch (e: Throwable) {
-        return
-    }
-}
-
-// 2 3 4 5 6 +8 9 10 11 8 13
