--- compiler/tests/org/jetbrains/kotlin/codegen/LineNumberTestGenerated.java
@@ -1,173 +1,0 @@
-/*
- * Copyright 2010-2020 JetBrains s.r.o. and Kotlin Programming Language contributors.
- * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
- */
-
-package org.jetbrains.kotlin.codegen;
-
-import com.intellij.testFramework.TestDataPath;
-import org.jetbrains.kotlin.test.JUnit3RunnerWithInners;
-import org.jetbrains.kotlin.test.KotlinTestUtils;
-import org.jetbrains.kotlin.test.TestMetadata;
-import org.junit.runner.RunWith;
-
-import java.io.File;
-import java.util.regex.Pattern;
-
-/** This class is generated by {@link org.jetbrains.kotlin.generators.tests.TestsPackage}. DO NOT MODIFY MANUALLY */
-@SuppressWarnings("all")
-@TestMetadata("compiler/testData/lineNumber")
-@TestDataPath("$PROJECT_ROOT")
-@RunWith(JUnit3RunnerWithInners.class)
-public class LineNumberTestGenerated extends AbstractLineNumberTest {
-    private void runTest(String testDataFilePath) throws Exception {
-        KotlinTestUtils.runTest(this::doTest, this, testDataFilePath);
-    }
-
-    public void testAllFilesPresentInLineNumber() throws Exception {
-        KotlinTestUtils.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("compiler/testData/lineNumber"), Pattern.compile("^(.+)\\.kt$"), null, true);
-    }
-
-    @TestMetadata("compiler/testData/lineNumber/custom")
-    @TestDataPath("$PROJECT_ROOT")
-    @RunWith(JUnit3RunnerWithInners.class)
-    public static class Custom extends AbstractLineNumberTest {
-        private void runTest(String testDataFilePath) throws Exception {
-            KotlinTestUtils.runTest(this::doTest, this, testDataFilePath);
-        }
-
-        public void testAllFilesPresentInCustom() throws Exception {
-            KotlinTestUtils.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("compiler/testData/lineNumber/custom"), Pattern.compile("^(.+)\\.kt$"), null, true);
-        }
-
-        @TestMetadata("beforeGotoToWhileStart.kt")
-        public void testBeforeGotoToWhileStart() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/beforeGotoToWhileStart.kt");
-        }
-
-        @TestMetadata("callWithCallInArguments.kt")
-        public void testCallWithCallInArguments() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/callWithCallInArguments.kt");
-        }
-
-        @TestMetadata("callWithReceiver.kt")
-        public void testCallWithReceiver() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/callWithReceiver.kt");
-        }
-
-        @TestMetadata("chainCall.kt")
-        public void testChainCall() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/chainCall.kt");
-        }
-
-        @TestMetadata("compileTimeConstant.kt")
-        public void testCompileTimeConstant() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/compileTimeConstant.kt");
-        }
-
-        @TestMetadata("functionCallWithDefault.kt")
-        public void testFunctionCallWithDefault() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/functionCallWithDefault.kt");
-        }
-
-        @TestMetadata("functionCallWithInlinedLambdaParam.kt")
-        public void testFunctionCallWithInlinedLambdaParam() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/functionCallWithInlinedLambdaParam.kt");
-        }
-
-        @TestMetadata("functionCallWithLambdaParam.kt")
-        public void testFunctionCallWithLambdaParam() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/functionCallWithLambdaParam.kt");
-        }
-
-        @TestMetadata("ifThen.kt")
-        public void testIfThen() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/ifThen.kt");
-        }
-
-        @TestMetadata("ifThenElse.kt")
-        public void testIfThenElse() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/ifThenElse.kt");
-        }
-
-        @TestMetadata("ifThenElseFalse.kt")
-        public void testIfThenElseFalse() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/ifThenElseFalse.kt");
-        }
-
-        @TestMetadata("inTheEndOfLambdaArgumentOfInlineCall.kt")
-        public void testInTheEndOfLambdaArgumentOfInlineCall() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/inTheEndOfLambdaArgumentOfInlineCall.kt");
-        }
-
-        @TestMetadata("initBlocks.kt")
-        public void testInitBlocks() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/initBlocks.kt");
-        }
-
-        @TestMetadata("multilineFunctionCall.kt")
-        public void testMultilineFunctionCall() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/multilineFunctionCall.kt");
-        }
-
-        @TestMetadata("multilineInfixCall.kt")
-        public void testMultilineInfixCall() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/multilineInfixCall.kt");
-        }
-
-        @TestMetadata("noParametersArgumentCallInExpression.kt")
-        public void testNoParametersArgumentCallInExpression() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/noParametersArgumentCallInExpression.kt");
-        }
-
-        @TestMetadata("primitiveNullChecks.kt")
-        public void testPrimitiveNullChecks() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/primitiveNullChecks.kt");
-        }
-
-        @TestMetadata("smapInlineAsArgument.kt")
-        public void testSmapInlineAsArgument() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/smapInlineAsArgument.kt");
-        }
-
-        @TestMetadata("smapInlineAsInfixArgument.kt")
-        public void testSmapInlineAsInfixArgument() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/smapInlineAsInfixArgument.kt");
-        }
-
-        @TestMetadata("smapInlineAsInlineArgument.kt")
-        public void testSmapInlineAsInlineArgument() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/smapInlineAsInlineArgument.kt");
-        }
-
-        @TestMetadata("smapInlineInIntrinsicArgument.kt")
-        public void testSmapInlineInIntrinsicArgument() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/smapInlineInIntrinsicArgument.kt");
-        }
-
-        @TestMetadata("tryCatchExpression.kt")
-        public void testTryCatchExpression() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/tryCatchExpression.kt");
-        }
-
-        @TestMetadata("tryCatchFinally.kt")
-        public void testTryCatchFinally() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/tryCatchFinally.kt");
-        }
-
-        @TestMetadata("tryFinally.kt")
-        public void testTryFinally() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/tryFinally.kt");
-        }
-
-        @TestMetadata("when.kt")
-        public void testWhen() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/when.kt");
-        }
-
-        @TestMetadata("whenSubject.kt")
-        public void testWhenSubject() throws Exception {
-            runTest("compiler/testData/lineNumber/custom/whenSubject.kt");
-        }
-    }
-}
