+++ compiler/tests/org/jetbrains/kotlin/generators/tests/GenerateCompilerTests.kt
@@ -391,10 +391,6 @@ fun main(args: Array<String>) {
             model("type/binding")
         }
 
-        testClass<AbstractLineNumberTest> {
-            model("lineNumber")
-        }
-
         testClass<AbstractSteppingTest>(useJunit4 = true) {
             model("debug/stepping", targetBackend = TargetBackend.JVM)
         }
@@ -483,10 +479,6 @@ fun main(args: Array<String>) {
                 targetBackend = TargetBackend.JVM_IR
             )
             model("loadJava/sourceJava", extension = "java", testMethod = "doTestSourceJava", targetBackend = TargetBackend.JVM_IR)
-        }
-
-        testClass<AbstractIrLineNumberTest> {
-            model("lineNumber", targetBackend = TargetBackend.JVM_IR)
         }
 
         testClass<AbstractIrSteppingTest>(useJunit4 = true) {
